+++ src/Providers/AddressesServiceProvider.php
@@ -1,0 +1,79 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rinvex\Addresses\Providers;
+
+use Illuminate\Support\ServiceProvider;
+use Rinvex\Addresses\Contracts\AddressContract;
+use Rinvex\Addresses\Console\Commands\MigrateCommand;
+
+class AddressesServiceProvider extends ServiceProvider
+{
+    /**
+     * The commands to be registered.
+     *
+     * @var array
+     */
+    protected $commands = [
+        MigrateCommand::class => 'command.rinvex.addresses.migrate',
+    ];
+
+    /**
+     * {@inheritdoc}
+     */
+    public function register()
+    {
+        // Merge config
+        $this->mergeConfigFrom(realpath(__DIR__.'/../../config/config.php'), 'rinvex.addresses');
+
+        // Bind eloquent models to IoC container
+        $this->app->singleton('rinvex.addresses.address', function ($app) {
+            return new $app['config']['rinvex.addresses.models.address']();
+        });
+        $this->app->alias('rinvex.addresses.address', AddressContract::class);
+
+        // Register console commands
+        ! $this->app->runningInConsole() || $this->registerCommands();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function boot()
+    {
+        // Load migrations
+        ! $this->app->runningInConsole() || $this->loadMigrationsFrom(__DIR__.'/../../database/migrations');
+
+        // Publish Resources
+        ! $this->app->runningInConsole() || $this->publishResources();
+    }
+
+    /**
+     * Publish resources.
+     *
+     * @return void
+     */
+    protected function publishResources()
+    {
+        $this->publishes([realpath(__DIR__.'/../../config/config.php') => config_path('rinvex.addresses.php')], 'rinvex-addresses-config');
+        $this->publishes([realpath(__DIR__.'/../../database/migrations') => database_path('migrations')], 'rinvex-addresses-migrations');
+    }
+
+    /**
+     * Register console commands.
+     *
+     * @return void
+     */
+    protected function registerCommands()
+    {
+        // Register artisan commands
+        foreach ($this->commands as $key => $value) {
+            $this->app->singleton($value, function ($app) use ($key) {
+                return new $key();
+            });
+        }
+
+        $this->commands(array_values($this->commands));
+    }
+}