+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-# Rinvex Addressable Change Log
+# Rinvex Addresses Change Log
 
 All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v0.0.2] - 2017-09-08
+- Fix many issues and apply many enhancements
+- Rename package rinvex/addresses from rinvex/addressable
+
 ## v0.0.1 - 2017-04-07
 - Commit first draft
+
+[v0.0.2]: https://github.com/rinvex/addresses/compare/v0.0.1...v0.0.2